test(fixture): only delete dist if exists should fix failure in CI

+++ test/fixtures/dist-dir/build.js
@@ -3,7 +3,9 @@ const path = require('path');
 
 const distPath = path.join(__dirname, 'npm');
 
-fs.rmdirSync(distPath, { recursive: true });
+if (fs.existsSync(distPath)) {
+    fs.rmdirSync(distPath, { recursive: true });
+}
 fs.mkdirSync(distPath, { recursive: true });
 fs.copyFileSync(require.resolve('./package.json'), path.join(distPath, 'package.json'));
 fs.writeFileSync(path.join(distPath, 'main.js'), `export default 'hello world';`);
